+++ CHANGELOG.md
@@ -3,3 +3,5 @@ v0.5.0
 
 * introduced `DateAfter` specification
 * introduced `Equal` that supports exact match for numbers, strings, dates and enums
+* introduced `In` that supports in operator for numbers, strings, dates and enums
+* deprecated `EqualEnum` (use `Equal` and `In`)